refactor: remove unused internal helpers and outdated duration parsing logic, reorganize wait-related compiler components, and add helper functions for member expressions
