+++ pkg/compiler/internal/context.go
@@ -12,58 +12,61 @@ import (
 	"github.com/MontFerret/ferret/v2/pkg/parser/diagnostics"
 )
 
-// ProgramContext holds state that lives for the entire compilation of one
-// source file. It is shared across the main body and all UDF compilations.
-//
-// Everything in ProgramContext is program-wide: writes from any function
-// compile (top-level body or a nested UDF body) are observable globally and
-// persist for the whole compilation. The final bytecode.Program reads its
-// metadata almost entirely from here.
-type ProgramContext struct {
-	Emitter             *core.Emitter
-	Constants           *core.ConstantPool
-	CatchTable          *core.CatchStack
-	UDFs                *core.UDFTable
-	HostParams          *core.HostParamTable
-	HostFunctions       *core.HostFunctionTable
-	Errors              *diagnostics.ErrorHandler
-	Source              *source.Source
-	UseAliases          map[string]string
-	aggregatePlanByHash map[uint64][]int
-	aggregatePlans      []*bytecode.AggregatePlan
-	OptimizationLevel   optimization.Level
-}
+type (
+	// ProgramContext holds state that lives for the entire compilation of one
+	// source file. It is shared across the main body and all UDF compilations.
+	//
+	// Everything in ProgramContext is program-wide: writes from any function
+	// compile (top-level body or a nested UDF body) are observable globally and
+	// persist for the whole compilation. The final bytecode.Program reads its
+	// metadata almost entirely from here.
+	ProgramContext struct {
+		Emitter             *core.Emitter
+		Constants           *core.ConstantPool
+		CatchTable          *core.CatchStack
+		UDFs                *core.UDFTable
+		HostParams          *core.HostParamTable
+		HostFunctions       *core.HostFunctionTable
+		Errors              *diagnostics.ErrorHandler
+		Source              *source.Source
+		UseAliases          map[string]string
+		aggregatePlanByHash map[uint64][]int
+		aggregatePlans      []*bytecode.AggregatePlan
+		OptimizationLevel   optimization.Level
+	}
 
-// FunctionContext holds state that is local to a single function body
-// compilation. A fresh FunctionContext is created for the top-level body
-// and for every UDF body.
-//
-// Function-local state is NEVER inherited from an enclosing function: a
-// nested UDF starts with an empty RegisterAllocator, an empty SymbolTable
-// (locals and globals), a fresh TypeTracker, a fresh LoopTable, and its
-// own UDFScope pointing at the UDF's body scope for inner UDF name
-// lookup. Carrying any of these over from a parent function would produce
-// wrong bytecode.
-//
-// Anything that must be visible across function boundaries (host params,
-// host function refs, constants, catch table, UDF metadata, the emitter)
-// lives on ProgramContext instead.
-type FunctionContext struct {
-	Registers *core.RegisterAllocator
-	Symbols   *core.SymbolTable
-	Types     *core.TypeTracker
-	Loops     *core.LoopTable
-	UDFScope  *core.UDFScope
-}
+	// FunctionContext holds state that is local to a single function body
+	// compilation. A fresh FunctionContext is created for the top-level body
+	// and for every UDF body.
+	//
+	// Function-local state is NEVER inherited from an enclosing function: a
+	// nested UDF starts with an empty RegisterAllocator, an empty SymbolTable
+	// (locals and globals), a fresh TypeTracker, a fresh LoopTable, and its
+	// own UDFScope pointing at the UDF's body scope for inner UDF name
+	// lookup. Carrying any of these over from a parent function would produce
+	// wrong bytecode.
+	//
+	// Anything that must be visible across function boundaries (host params,
+	// host function refs, constants, catch table, UDF metadata, the emitter)
+	// lives on ProgramContext instead.
+	FunctionContext struct {
+		Registers *core.RegisterAllocator
+		Symbols   *core.SymbolTable
+		Types     *core.TypeTracker
+		Loops     *core.LoopTable
+		UDFScope  *core.UDFScope
+	}
 
-// CompilationSession is the thin coordinator passed to all compilers.
-// It provides access to both program-wide and function-local state.
-// The Function pointer is swapped by withFunctionCompileState during UDF
-// compilation; the Program pointer never changes during a single Compile.
-type CompilationSession struct {
-	Program  *ProgramContext
-	Function *FunctionContext
-}
+	// CompilationSession is the thin coordinator passed to all compilers.
+	// It provides access to both program-wide and function-local state.
+	// The Function pointer is swapped by withFunctionCompileState during UDF
+	// compilation; the Program pointer never changes during a single Compile.
+
+	CompilationSession struct {
+		Program  *ProgramContext
+		Function *FunctionContext
+	}
+)
 
 // NewFunctionContext creates a fresh function-local compilation state.
 func NewFunctionContext(constants *core.ConstantPool) *FunctionContext {
@@ -107,11 +110,11 @@ func (c *ProgramContext) AddAggregatePlan(plan *bytecode.AggregatePlan) int {
 		return -1
 	}
 
-	hash := aggregatePlanHash(plan)
+	hash := c.aggregatePlanHash(plan)
 
 	if existing, ok := c.aggregatePlanByHash[hash]; ok {
 		for _, idx := range existing {
-			if idx >= 0 && idx < len(c.aggregatePlans) && areAggregatePlansEqual(c.aggregatePlans[idx], plan) {
+			if idx >= 0 && idx < len(c.aggregatePlans) && c.areAggregatePlansEqual(c.aggregatePlans[idx], plan) {
 				return idx
 			}
 		}
@@ -136,7 +139,7 @@ func (c *ProgramContext) AggregatePlans() []bytecode.AggregatePlan {
 	return plans
 }
 
-func areAggregatePlansEqual(a, b *bytecode.AggregatePlan) bool {
+func (c *ProgramContext) areAggregatePlansEqual(a, b *bytecode.AggregatePlan) bool {
 	if len(a.Keys) == 0 || len(b.Keys) == 0 {
 		return a == b
 	}
@@ -154,7 +157,7 @@ func areAggregatePlansEqual(a, b *bytecode.AggregatePlan) bool {
 	return true
 }
 
-func aggregatePlanHash(plan *bytecode.AggregatePlan) uint64 {
+func (c *ProgramContext) aggregatePlanHash(plan *bytecode.AggregatePlan) uint64 {
 	h := fnv.New64a()
 	h.Write([]byte("aggregate_plan:"))
