+++ pkg/compiler/internal/wait_predicate.go
@@ -13,19 +13,28 @@ import (
 	"github.com/MontFerret/ferret/v2/pkg/runtime"
 )
 
-type waitForPredicateMode int
+type (
+	waitForPredicateMode int
 
-type waitPredicateCompileConfig struct {
-	predExpr      fql.IExpressionContext
-	jitterLiteral *float64
-	mode          waitForPredicateMode
-	timeoutReg    bytecode.Operand
-	everyReg      bytecode.Operand
-	capEveryReg   bytecode.Operand
-	backoff       core.RetryBackoff
-	jitterReg     bytecode.Operand
-	hasJitter     bool
-}
+	waitPredicateCompileConfig struct {
+		predExpr      fql.IExpressionContext
+		jitterLiteral *float64
+		mode          waitForPredicateMode
+		timeoutReg    bytecode.Operand
+		everyReg      bytecode.Operand
+		capEveryReg   bytecode.Operand
+		backoff       core.RetryBackoff
+		jitterReg     bytecode.Operand
+		hasJitter     bool
+	}
+)
+
+const (
+	waitForPredicateModeBool waitForPredicateMode = iota
+	waitForPredicateModeExists
+	waitForPredicateModeNotExists
+	waitForPredicateModeValue
+)
 
 func (c *WaitCompiler) compilePredicate(ctx fql.IWaitForPredicateExpressionContext) bytecode.Operand {
 	config, ok := c.prepareWaitPredicateConfig(ctx)
@@ -81,76 +90,6 @@ func (c *WaitCompiler) prepareWaitPredicateConfig(ctx fql.IWaitForPredicateExpre
 	c.normalizeWaitPredicateConfig(&config)
 
 	return config, true
-}
-
-func legacyWaitForOrThrowNode(expr fql.IExpressionContext) antlr.ParserRuleContext {
-	if expr == nil || expr.LogicalOrOperator() == nil {
-		return nil
-	}
-
-	return bareThrowExpressionNode(expr.GetRight())
-}
-
-func bareThrowExpressionNode(expr fql.IExpressionContext) antlr.ParserRuleContext {
-	if expr == nil {
-		return nil
-	}
-
-	if expr.UnaryOperator() != nil || expr.LogicalAndOperator() != nil || expr.LogicalOrOperator() != nil || expr.GetTernaryOperator() != nil {
-		return nil
-	}
-
-	return bareThrowPredicateNode(expr.Predicate())
-}
-
-func bareThrowPredicateNode(pred fql.IPredicateContext) antlr.ParserRuleContext {
-	if pred == nil {
-		return nil
-	}
-
-	if pred.EqualityOperator() != nil || pred.ArrayOperator() != nil || pred.InOperator() != nil || pred.LikeOperator() != nil {
-		return nil
-	}
-
-	return bareThrowAtomNode(pred.ExpressionAtom())
-}
-
-func bareThrowAtomNode(atom fql.IExpressionAtomContext) antlr.ParserRuleContext {
-	if atom == nil {
-		return nil
-	}
-
-	if atom.MultiplicativeOperator() != nil || atom.AdditiveOperator() != nil || atom.RegexpOperator() != nil {
-		return nil
-	}
-
-	variable := atom.Variable()
-	if variable == nil || !strings.EqualFold(matchVariableName(variable), "THROW") {
-		return nil
-	}
-
-	node, ok := variable.(antlr.ParserRuleContext)
-	if !ok {
-		return nil
-	}
-
-	return node
-}
-
-func resolveWaitPredicateMode(hasValue, hasExists, hasNot bool) waitForPredicateMode {
-	if hasValue {
-		return waitForPredicateModeValue
-	}
-
-	if hasExists {
-		if hasNot {
-			return waitForPredicateModeNotExists
-		}
-
-		return waitForPredicateModeExists
-	}
-
-	return waitForPredicateModeBool
 }
 
 func (c *WaitCompiler) buildWaitPredicateConfig(
@@ -333,86 +272,3 @@ func (c *WaitCompiler) compileBackoffClause(ctx fql.IBackoffClauseContext) core.
 		return core.RetryBackoffNone
 	}
 }
-
-func literalFromExpression(ctx fql.IExpressionContext) fql.ILiteralContext {
-	if ctx == nil {
-		return nil
-	}
-
-	predicate := ctx.Predicate()
-	if predicate == nil {
-		return nil
-	}
-
-	atom := predicate.ExpressionAtom()
-	if atom == nil {
-		return nil
-	}
-
-	return atom.Literal()
-}
-
-func literalExistsFromExpression(ctx fql.IExpressionContext) (bool, bool) {
-	lit := literalFromExpression(ctx)
-	if lit == nil {
-		return false, false
-	}
-
-	switch {
-	case lit.NoneLiteral() != nil:
-		return false, true
-	case lit.StringLiteral() != nil:
-		if str, ok := parseStringLiteralConst(lit.StringLiteral()); ok {
-			return str.String() != "", true
-		}
-		return false, false
-	case lit.ArrayLiteral() != nil:
-		arr := lit.ArrayLiteral()
-		return arr.ArgumentList() != nil, true
-	case lit.ObjectLiteral() != nil:
-		obj := lit.ObjectLiteral()
-		return len(obj.AllPropertyAssignment()) > 0, true
-	default:
-		return true, true
-	}
-}
-
-func literalTruthinessFromExpression(ctx fql.IExpressionContext) (bool, bool) {
-	lit := literalFromExpression(ctx)
-	if lit == nil {
-		return false, false
-	}
-
-	switch {
-	case lit.NoneLiteral() != nil:
-		return false, true
-	case lit.BooleanLiteral() != nil:
-		return strings.ToLower(lit.BooleanLiteral().GetText()) == "true", true
-	case lit.IntegerLiteral() != nil:
-		val, err := strconv.Atoi(lit.IntegerLiteral().GetText())
-		if err != nil {
-			return false, false
-		}
-		return val != 0, true
-	case lit.FloatLiteral() != nil:
-		val, err := strconv.ParseFloat(lit.FloatLiteral().GetText(), 64)
-		if err != nil {
-			return false, false
-		}
-		return val != 0, true
-	case lit.StringLiteral() != nil:
-		if str, ok := parseStringLiteralConst(lit.StringLiteral()); ok {
-			return str.String() != "", true
-		}
-		return false, false
-	default:
-		return true, true
-	}
-}
-
-const (
-	waitForPredicateModeBool waitForPredicateMode = iota
-	waitForPredicateModeExists
-	waitForPredicateModeNotExists
-	waitForPredicateModeValue
-)
